+++ src/controller/series.controller.ts
@@ -1,17 +1,20 @@
+import { AccessJwtPayload } from '@/interface';
 import { SeriesDto } from '@/model';
-import { SeriesService } from '@/service';
+import { SeriesService, AuthService } from '@/service';
 import { checkIsNumber } from '@/utils/global-func';
 import {
   BadRequestException,
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   NotFoundException,
   Param,
   Post,
   Put,
   Query,
+  Req,
   UseGuards,
 } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
@@ -19,7 +22,10 @@ import { ApiTags } from '@nestjs/swagger';
 
 @Controller('series')
 export class SeriesController {
-  constructor(private readonly seriesService: SeriesService) {}
+  constructor(
+    private readonly seriesService: SeriesService,
+    private readonly authService: AuthService,
+  ) {}
 
   @Get()
   @ApiTags('content-series')
@@ -54,7 +60,21 @@ export class SeriesController {
   @Post()
   @ApiTags('content-series')
   @UseGuards(AuthGuard('jwt-access'))
-  async createSeries(@Body() body: SeriesDto) {
+  async createSeries(@Body() body: SeriesDto, @Req() req) {
+    const jwtPayload: AccessJwtPayload = req.user;
+
+    const member = await this.authService.memberById(jwtPayload._id);
+
+    if (!Boolean(member)) {
+      throw new BadRequestException('Thành viên không tồn tại!.');
+    }
+
+    if (member.role === 'member') {
+      throw new ForbiddenException(
+        'Bạn không có quyền thao tác với chuỗi bài viết!.',
+      );
+    }
+
     const isExist = await this.seriesService.checkNameExist(body.title);
 
     if (isExist) {
@@ -67,11 +87,35 @@ export class SeriesController {
   @Put(':id')
   @ApiTags('content-series')
   @UseGuards(AuthGuard('jwt-access'))
-  async updateSeries(@Body() body: SeriesDto, @Param('id') id: string) {
+  async updateSeries(
+    @Body() body: SeriesDto,
+    @Param('id') id: string,
+    @Req() req,
+  ) {
+    const jwtPayload: AccessJwtPayload = req.user;
+
+    const member = await this.authService.memberById(jwtPayload._id);
+
+    if (!Boolean(member)) {
+      throw new BadRequestException('Thành viên không tồn tại!.');
+    }
+
+    if (member.role === 'member') {
+      throw new ForbiddenException(
+        'Bạn không có quyền thao tác với chuỗi bài viết!',
+      );
+    }
+
     const series = await this.seriesService.getSeriesById(id);
 
     if (!series) {
-      throw new BadRequestException('Thể loại cần chỉnh sửa không tồn tại.!');
+      throw new BadRequestException('Thể loại cần chỉnh sửa không tồn tại!');
+    }
+
+    if (member._id !== series.created_by._id) {
+      throw new ForbiddenException(
+        'Bạn không phải là người tạo của chuỗi bài viết này!',
+      );
     }
 
     return this.seriesService.update(series._id, body);
@@ -80,11 +124,31 @@ export class SeriesController {
   @Delete(':id')
   @ApiTags('content-series')
   @UseGuards(AuthGuard('jwt-access'))
-  async deleteSeries(@Param('id') id: string) {
+  async deleteSeries(@Param('id') id: string, @Req() req) {
+    const jwtPayload: AccessJwtPayload = req.user;
+
+    const member = await this.authService.memberById(jwtPayload._id);
+
+    if (!Boolean(member)) {
+      throw new BadRequestException('Thành viên không tồn tại!.');
+    }
+
+    if (member.role === 'member') {
+      throw new ForbiddenException(
+        'Bạn không có quyền thao tác với chuỗi bài viết!.',
+      );
+    }
+
     const series = await this.seriesService.getSeriesById(id);
 
     if (!series) {
       throw new BadRequestException('Thể loại cần xoá không tồn tại.!');
+    }
+
+    if (member._id !== series.created_by._id) {
+      throw new ForbiddenException(
+        'Bạn không phải là người tạo chuỗi bài viết này!.',
+      );
     }
 
     return await this.seriesService.delete(id);
